Change argument-list type
If `Tuple[str, ...]` is used, then each individual positional argument
must be a tuple of strings.  See PEP-484.

`Any` is used instead, to allow for `float`s in the accompanying unit
test, and `dict`s in future applications.

With these changes, `/mix_utils` passes `mypy`.

References:
* https://peps.python.org/pep-0484/#arbitrary-argument-lists-and-default-argument-values

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>