Add type annotations
With these annotations, the following command (with the package
`types-pytz` also installed) reports no issues:

    mypy case_mapping example.py tests

Adding type review to unit testing is left for a future patch.

No effects were observed on Make-managed files.

References:
* https://docs.python.org/3/library/typing.html#the-type-of-class-objects

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>