Explicitly add Facet typing
This patch satisfies `case_validate` requirements by use of explicit
typing in the data graph, rather than relying on an external ontology
definition.  The realized code suggests another pattern relying on the
Python class hierarchy should be followed, but that is left for future
work.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>