Move drafting:fileLocalPath to PathRelationshipFacet, updating class hierarchy
A prior patch in this series tried updating a non-existent property
`uco-observable:fileLocalPath` to a `FileFacet` attached to the subject
file within the SD card.  This highlighted that the property is not
defined in the ontology, and this patch was written to describe why it
should not be.

"Local" is a confusing term, due to absence of context - to what
directory is this path local?  Is that directory through the user
interface on the phone, which would confuse in details from a running
operating system?  Or is the directory local to the extraction computer,
which would confuse in details from the investigation?

This patch removes the ambiguity inherent in `drafting:fileLocalPath`
and moves the path to a `PathRelationFacet`, which attaches to an
`ObservableRelationship` that can define the basis of "local".

This patch adds a guess that the path's new spelling suggested an
extraction directory special to the investigation.  Chain of custody
details are left for a future patch to implement.

A side-effect of this patch is that some of the API usage was found to
not make use of the subclass hierarchy.  `super()` and anonymous
argument & parameter forwarding is now used to delegate handling generic
properties to their originating superclasses.  E.g., note how a
`uco-observable:File` picked up a `core:description` property by the
property's addition as a keyword parameter on `base.ObjectEntity`.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>