Update InvestigativeAction to follow ontology's class hierarchy
`uco-action:Action` did not yet have a class implemented in `action.py`.
All of the object-management work---aside from the `@type`
designation---that occurred in `class InvestigativeAction` was actually
work that is supposed to be done by `class Action`.

Separately, `uco-action:ActionReferencesFacet` was removed in UCO 0.8.0,
its functionality collapsed into `uco-action:Action`.

This patch adapts `class InvestigativeAction` to follow
inheritance-based code.  There is potential to adapt the class to handle
work involving generated `case-investigation:ProvenanceRecord`
logistics, but that is left for future work.

Function signatures' keyword parameters have been alphabetized.

The keyword parameters for `uco-action:object` and `uco-action:result`
are left in their plural spelling because `object` is a keyword in
Python.

A follow-on patch will regenerate Make-managed files.

References:
* https://unifiedcyberontology.org/releases/0.8.0/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>