Align ApplicationVersion as a UcoInherentCharacterizationThing
ApplicationVersion was being treated as a UcoObject, and being appended
to the Bundle.  The `_append_stuff` method instead needed to be called
with `objects=True` in order to inline the nodes.

Alternatively, the generated JSON-LD object could be given a `"@graph"`
key, instead of housing everything in a Bundle.  This seems like it will
be necessary by the time shared object-references are added, e.g. with
`uco-types:Hash`.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>