Add steps to build project in publishing workflow
Workflow adjustments are drawn from parts of a blog post.

References:
* https://johnfraney.ca/blog/how-to-publish-a-python-package-with-poetry-and-github-actions/

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>