Insert intermediary Compilation classes between Investigation and UcoObject
One effect reflected back into the example script is that `Compilation`,
`ContextualCompilation`, and their subclasses now must be initialized
with at least one object `UcoObject` linked in the `core_objects`
keyword parameter.  Documentation is added to note this is to guarantee
minimum cardinality constraints in the ontology are always satisfied.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>